backend/requirements.txt
python-dotenv
openai
Flask